Add sentry on errors. (#264)
Fixes #7 

After this lands, we have a bit of paperwork to do. We need to create 2 Sentry projects. (We already have https://sentry.prod.mozaws.net/operations/rs-uptake-health-dev/ but I don't like that name any more) and we need to ask Ops to set SENTRY_DSN on ALL Lambda configs.